fix issue #1 - `is_link` used as intended but assigned as `is_pe`
the condition is unchanged but branches are swapped for clarity

The signatures of PE (portable executable) images (such as .sys or .exe files)
are encoded differently in the resulting .cat file. Due to a logic
bug PE files signature were encoded as link (and non-PE as images).

This patch fixes this.