Fixed calling generate-cat-file
When you put something in double quotes it will be
passed as one parameter (one argv) to the program.
This is a way to pass whitespace to programs. In
our case the double quotes were wrong since we
want to pass -G and the GUID (or -T and the timestamp)
as seperate (argv) parameters.

Now it builds with WinDRBD and the CAT files generated
can be used to load the driver.